rename loader to interfaces
