+++ src/lr_reduction/io/__init__.py
@@ -1,4 +1,4 @@
-from lr_reduction.io.loader import ConfigLoaderInterface, Loader, RunLoaderInterface
+from lr_reduction.io.interfaces import ConfigLoaderInterface, Loader, RunLoaderInterface
 from lr_reduction.io.run_data import RunData
 
 __all__ = [
